Add files via upload
"Introducing 'Guess the Number' – a Java-based game where players must guess a number between 1 and 100. With each incorrect guess, the game guides you by suggesting whether the number is higher or lower. At the end of the game, your performance is scored, offering a fun way to track your guessing skills. Fully coded in Java, this game is a perfect blend of logic and entertainment